make sweepers_driver main routine importable
